added playGame and result function plus commands
created playGame function using while loop to be able to
play 5 rounds of the game and moved all otherfunctions into it.
added necessery constants and called on the functions to make
whole game play smoothly. playGame asks 5 times for the choice
of hand and after each round logs the results to console.
At the end outcome/result of the game is announced.

One concerne to be resolved is when choice of hand is misspelled
player is notified/alert that hand is not recognised and awarded
no points for that round and playGame function moves to next round.