capital hand call, comment out func & moved script
1. changed to all capital the name of the hands for addEventListener
so that when button pressed the playRound() can read it and compare it
without an issue.
2.commented out the function get human choice as we will not use
prompt to get human to play the game but buttons instead.
3. I was getting typeError for addEventListener, cannot read null.
after quick google search I have found and inplemented change to <script>
and decided to move it from <head> to end of the body as: "This is probably
because the script is executed before the page loads. By placing the script
at the bottom of the page, I circumvented the problem."
Personally not sure if it is a best practice but it has solved my issue.